Audio: Fix Ali customer issues
1) Fix ANR when no audio-in data during recording

Timeout for reading from audio-in socket was a huge number.
Due to the large timeout value, Audio VHAL blocks in in_read()
until audio-in data is available. If there is no audio-in data
supplied, then this long wait results into ANR in the system.

The proposed fix is to set 500ms as the audio-in socket read timeout.
With this, Audio VHAL in_read() blocks at most for 500ms and returns
to framework with null data.

2) Sometimes sound recorder crashes

Avoid freeing ssi.in which corrupts stream_in of framework. This fixes
crash in audioserver.

3) Sometimes AUDIO_IN command is sent on cg_proxy start up

Set ssi to null unconditionally, without checking client connection.

Change-Id: Ic6b12977062894fb4e053be995cc9db1ec6ae7fb
Tracked-On: OAM-92229
Tracked-On: OAM-92020
Tracked-On: OAM-92268
Signed-off-by: Shakthi Prashanth M<shakthi.prashanth.m@intel.com>