Camera: Fix for Freeze and Green flicker issue in WKSZ app
wksz app requests to process for both Preview
and Capture streams simultaneously. Due to this
reason, Camera vHAL should handle both the streams
simultaneously with the existing Inflight Queue,
but by default it can handle only one in the
existing queue logic. Hence Camera vHAL is not able
to handle the streams after some time and obviously
time out happens. Due to this error return to the
framework and app, we are observing Camera freeze
in the preview and that is ultimately causing the
Green flicker issue after re-launch the app.

The fix for this issue is to increase the Queue
size to four and hence it can handle two streams
simultaneously.

Change-Id: I1815c11b8c5b15ae8c99e56ffd69b2bd6dd7888a
Tracked-On: OAM-95465
Signed-off-by: Muhammad Aksar <muhammad.aksar@intel.com>