[W/A] Add additional check to fix color reversal issue
As the buffer format, usage for capture stream is same for
camera2, wksz and Rura app, fixing color reversal for Rura
app introduces color reversal issue in Camera2 and wksz app.

Fix the issue by checking for number of configured streams
in additional to usage flag to determine the format.

Tracked-On: VVP0-2813
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>