Fix for Camera error issue when random close and open
It fixes the Camera close issue when do random
close and open camera sessions especially while
refresh the client browsers.

Fix introduces a timeout mechanism which waits
for max 12ms time for camera close session, and
if it is not happening within that time, will
break the loop forcefully to return closed state
to Android framework. Otherwise Camera service
will not get lock for the next Camera client.

This logic ensures the lock for all the Camera
clients in any abnormal condition to make the
Camera System more stable and reliable.

Tracked-On: OAM-96874
Signed-off-by: Muhammad Aksar <muhammad.aksar@intel.com>